Histogram elementwise; optionally accumulate
Factor out histogramming code (histogramming.hpp) from processors
(histogram.hpp, histogram_elementwise.hpp). Rewrite single histogram
processor accordingly. Remove "batched" single histograms (for now) as
they are redundant with elementwise with single element.

Discard separate processor for bin-increment-batch journaling (hide
journaling inside histogramming classes to whose state it is tightly
coupled).

Some unrelated changes that accompany the addition of -Wconversion.

+++ include/flimevt/bh_spc.hpp
@@ -57,7 +57,7 @@ struct bh_spc_event {
         // The documentation somewhat confusingly says that these bits are
         // "inverted", but what they mean is that the TTL inputs are active
         // low. The bits in the FIFO data are not inverted.
-        return unsigned(bytes[1]) >> 4;
+        return static_cast<std::uint8_t>(unsigned(bytes[1]) >> 4);
     }
 
     /**
